pyproject.toml
[tool.poetry]
name = "fastapi-lambda"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.62.0"
mangum = "^0.10.0"
aws-lambda-powertools = "^1.8.0"
pydantic = "^1.7.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
uvicorn = "^0.12.3"
invoke = "^1.4.1"
httpx = "^0.16.1"
"aws-cdk.core" = "^1.76.0"
"aws-cdk.aws-lambda" = "^1.76.0"
"aws-cdk.aws-ecr" = "^1.76.0"
"aws-cdk.aws-apigatewayv2-integrations" = "^1.76.0"
black = "^20.8b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
